Make platform.mac_ver() return '10.10' on Yosemite.
Previous wrong behaviour was:
$ uname -r
> 14.0.0
$ buildout.python/python-2.4/bin/python -c 'print __import__("platform").mac_ver()'
> ('10.9', ('', '', ''), 'x86_64')

New correct behaviour is:
$ buildout.python/python-2.4/bin/python -c 'print __import__("platform").mac_ver()'
> ('10.10', ('', '', ''), 'x86_64')

I've backported the _mac_ver_xml() function from Python 2.7 to Python 2.4,
which reads /System/Library/CoreServices/SystemVersion.plist to get the
current version number instead of using the Gestalt API, because the
Gestalt API has a really awesome bug in it wherein it returns 0x1080 on
OS X 10.8, 0x1090 on OS X 10.9, and 0x1090 on OS X 10.10.

This bug was manifesting in a buildout failure when I tried to get
buildout to build PIL on Yosemite, because bdist_egg in
setuptools.command was generating an egg with the name
"PIL-1.1.6-py2.4-macosx-10.10-x86_64.egg" (as it gets its idea of
OSX's version number from $PREFIX/lib/python2.4/config/Makefile)
whereas setuptools.command.easy_install was looking for an egg with
the name "PIL-1.1.6-py2.4-macosx-10.9-x86_64.egg" (because it gets its
idea of OSX's version number from Gestalt) and subsequently raising an
exception like:

 "No eggs found in /var/folders/nd/nz2j950x6gd51bjfj2316y580000gr/T/easy_install-gkyUY8/PIL-1.1.6/egg-dist-tmp-Rkpvi3 (setup script problem?)"

Aaa.