Add `--enable-optimizations` configure flag.
This enables profile guided optimizations (PGO) on supported Python
versions. PGO increases build time significantly, but results in
Python being faster at runtime.

The configure flag is backported to the 2.7, 3.5 and 3.6 branches,
but only available in the last 3.6 release as of yet. Once 2.7.13 and
3.5.3 are released, it will take effect for those as well.

See for example http://bugs.python.org/issue26359.